Image-Captioning-main (2)/requirements.txt
# =============================================================================
# MULTILINGUAL IMAGE CAPTIONING - REQUIREMENTS
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core AI/ML Libraries
# -----------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# -----------------------------------------------------------------------------
# Image Processing
# -----------------------------------------------------------------------------
Pillow>=10.0.0
numpy<2.0.0
opencv-python>=4.8.0
scikit-image>=0.19.0

# -----------------------------------------------------------------------------
# Web Interface
# -----------------------------------------------------------------------------
gradio>=4.0.0,<6.0.0

# -----------------------------------------------------------------------------
# Natural Voice (Text-to-Speech)
# -----------------------------------------------------------------------------
edge-tts>=6.1.0
pygame>=2.5.0
pyttsx3>=2.90

# -----------------------------------------------------------------------------
# OCR - Text Detection in Images
# -----------------------------------------------------------------------------
easyocr>=1.7.0

# -----------------------------------------------------------------------------
# Translation & Language Detection
# -----------------------------------------------------------------------------
deep-translator>=1.11.0
langdetect>=1.0.9

# -----------------------------------------------------------------------------
# NLP & Text Processing
# -----------------------------------------------------------------------------
nltk>=3.7

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
tqdm>=4.66.0
requests>=2.28.0
matplotlib>=3.5.2

# -----------------------------------------------------------------------------
# Original Project Dependencies (for backward compatibility)
# -----------------------------------------------------------------------------
pycocotools>=2.0.4
jupyter>=1.0.0
